Refactor: major code reorganization and module extraction
